+++ src/addons/send2ue/release_notes.md
@@ -1,9 +1,9 @@
 ## Patch Changes
-* Lowered aggressiveness of socket names
-  * [173](https://github.com/poly-hammer/BlenderTools/issues/173)
+* Fixed addon preferences setter/getters to be compatible with Blender 5.0
+  * [174](https://github.com/poly-hammer/BlenderTools/issues/174)
 
 ## Special Thanks
-@Daerst
+@kelitraynaud
 
 ## Tests Passing On
 * Blender `3.6`, `4.2` (installed from blender.org)